+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Notable changes to this project are documented herein.
 ### Added
 - Support for Neo4j 3.1.
 - Support for Python 3.6.
+- New Cypher console
 
 ### Changed
 - Official driver 1.1.0 dependency replaces embedded 1.0 series driver. This is the first (and so far only) project dependency.